wip: Class polymorphism for Storage
