+++ cpp/arcticdb/storage/azure/azure_storage.hpp
@@ -23,11 +23,7 @@
 
 namespace arcticdb::storage::azure {
 
-class AzureStorage final : public Storage<AzureStorage> {
-
-    using Parent = Storage<AzureStorage>;
-    friend Parent;
-
+class AzureStorage final : public Storage {
   public:
     // friend class AzureTestClientAccessor<AzureStorage>;
     using Config = arcticdb::proto::azure_storage::Config;
@@ -40,25 +36,27 @@ class AzureStorage final : public Storage<AzureStorage> {
     std::string get_key_path(const VariantKey& key) const;
 
   protected:
-    void do_write(Composite<KeySegmentPair>&& kvs);
+    void do_write(Composite<KeySegmentPair>&& kvs) final;
 
-    void do_update(Composite<KeySegmentPair>&& kvs, UpdateOpts opts);
+    void do_update(Composite<KeySegmentPair>&& kvs, UpdateOpts opts) final;
 
-    void do_read(Composite<VariantKey>&& ks, const ReadVisitor& visitor, ReadKeyOpts opts);
+    void do_read(Composite<VariantKey>&& ks, const ReadVisitor& visitor, ReadKeyOpts opts) final;
 
-    void do_remove(Composite<VariantKey>&& ks, RemoveOpts opts);
+    void do_remove(Composite<VariantKey>&& ks, RemoveOpts opts) final;
 
-    void do_iterate_type(KeyType key_type, const IterateTypeVisitor& visitor, const std::string &prefix);
+    void do_iterate_type(KeyType key_type, const IterateTypeVisitor& visitor, const std::string &prefix) final;
 
-    bool do_key_exists(const VariantKey& key);
+    bool do_key_exists(const VariantKey& key) final;
 
-    bool do_supports_prefix_matching() {
+    bool do_supports_prefix_matching() final {
         return true;
     }
 
-    bool do_fast_delete() {
+    bool do_fast_delete() final {
         return false;
     }
+
+    std::string do_storage_specific(const VariantKey&) final { return {}; };
 
   private:
     Azure::Storage::Blobs::BlobContainerClient container_client_;
